Add zeros before adding space
Adding the space before adding the zero resulted in the "fun" behaviour
of getting a result like "0123 45". So instead of getting a nice
grouping of 3, we get one grouping of 4 and one of 2. This should
(finally) be fixed now. Don´t really know what took me so long to fix
it since it has been bothering me since the first day I saw it...

Anyhow, fixed it.